LayoutTests/css3/scroll-snap/scroll-snap-inherit.html

<!DOCTYPE html>
<html>
    <head>
        <style>
            .topLevel {
                width: 400px;
                -webkit-scroll-snap-type: proximity;
                -webkit-scroll-snap-points-x: repeat(200%);
                -webkit-scroll-snap-points-y: repeat(200%);
                -webkit-scroll-snap-destination: 25% 25%;
                -webkit-scroll-snap-coordinate: 20px 20px;
            }
            .horizontalGallery {
                width: 400px;
                height: 50px;
                overflow-y: hidden;
                overflow-x: auto;
            }
            .horizontalGalleryDrawer {
                width: 3000px;
                height: 50px;
            }
            .colorBox {
                height: 50px;
                width: 400px;
                float: left;
            }

            .noInherit {
                -webkit-scroll-snap-type: mandatory;
                -webkit-scroll-snap-points-x: repeat(100%);
                -webkit-scroll-snap-points-y: repeat(100%);
                -webkit-scroll-snap-destination: 10% 11%;
                -webkit-scroll-snap-coordinate: none;
            }

            .inheritType {
                -webkit-scroll-snap-type: inherit;
                -webkit-scroll-snap-points-x: repeat(100%);
                -webkit-scroll-snap-points-y: repeat(100%);
                -webkit-scroll-snap-destination: 0px 0px;
                -webkit-scroll-snap-coordinate: 15px 122px;
            }

            .inheritXPoints {
                -webkit-scroll-snap-type: mandatory;
                -webkit-scroll-snap-points-x: inherit;
                -webkit-scroll-snap-points-y: repeat(100%);
                -webkit-scroll-snap-destination: 50% 70%;
                -webkit-scroll-snap-coordinate: 100% 50%;
            }

            .inheritYPoints {
                -webkit-scroll-snap-type: mandatory;
                -webkit-scroll-snap-points-x: repeat(100%);
                -webkit-scroll-snap-points-y: inherit;
                -webkit-scroll-snap-destination: 10px 20px;
                -webkit-scroll-snap-coordinate: 50% 100%;
            }

            .inheritDestination {
                -webkit-scroll-snap-type: proximity;
                -webkit-scroll-snap-points-x: repeat(100%);
                -webkit-scroll-snap-points-y: repeat(100%);
                -webkit-scroll-snap-destination: inherit;
                -webkit-scroll-snap-coordinate: 10px 10px;
            }

            .inheritSnapCoordinate {
                -webkit-scroll-snap-type: mandatory;
                -webkit-scroll-snap-points-x: repeat(100%);
                -webkit-scroll-snap-points-y: repeat(100%);
                -webkit-scroll-snap-destination: 10% 10%;
                -webkit-scroll-snap-coordinate: inherit;
            }

            .inheritAll {
                -webkit-scroll-snap-type: inherit;
                -webkit-scroll-snap-points-x: inherit;
                -webkit-scroll-snap-points-y: inherit;
                -webkit-scroll-snap-destination: inherit;
                -webkit-scroll-snap-coordinate: inherit
            }

            .inheritNone {
            }

            #item0 { background-color: red; }
            #item1 { background-color: green; }
            #item2 { background-color: blue; }
            #item3 { background-color: aqua; }
            #item4 { background-color: yellow; }
            #item5 { background-color: fuchsia; }
        </style>
        <script src="../../resources/js-test-pre.js"></script>
    </head>
    <body>
        <div class="topLevel">
            <div id="noInherit" class="horizontalGallery noInherit">
                <div class="horizontalGalleryDrawer">
                    <div id="item0" class="colorBox"></div>
                    <div id="item1" class="colorBox"></div>
                    <div id="item2" class="colorBox"></div>
                    <div id="item3" class="colorBox"></div>
                    <div id="item4" class="colorBox"></div>
                    <div id="item5" class="colorBox"></div>
                </div>
            </div>
            <div id="inheritType" class="horizontalGallery inheritType" style="-webkit-scroll-snap-type: inherit;">
                <div class="horizontalGalleryDrawer">
                    <div id="item0" class="colorBox"></div>
                    <div id="item1" class="colorBox"></div>
                    <div id="item2" class="colorBox"></div>
                    <div id="item3" class="colorBox"></div>
                    <div id="item4" class="colorBox"></div>
                    <div id="item5" class="colorBox"></div>
                </div>
            </div>
            <div id="inheritXPoints" class="horizontalGallery inheritXPoints" style="-webkit-scroll-snap-points-x: inherit;">
                <div class="horizontalGalleryDrawer">
                    <div id="item0" class="colorBox"></div>
                    <div id="item1" class="colorBox"></div>
                    <div id="item2" class="colorBox"></div>
                    <div id="item3" class="colorBox"></div>
                    <div id="item4" class="colorBox"></div>
                    <div id="item5" class="colorBox"></div>
                </div>
            </div>
            <div id="inheritYPoints" class="horizontalGallery inheritYPoints" style="-webkit-scroll-snap-points-y: inherit;">
                <div class="horizontalGalleryDrawer">
                    <div id="item0" class="colorBox"></div>
                    <div id="item1" class="colorBox"></div>
                    <div id="item2" class="colorBox"></div>
                    <div id="item3" class="colorBox"></div>
                    <div id="item4" class="colorBox"></div>
                    <div id="item5" class="colorBox"></div>
                </div>
            </div>
            <div id="inheritDestination" class="horizontalGallery inheritDestination" style="-webkit-scroll-snap-destination: inherit;">
                <div class="horizontalGalleryDrawer">
                    <div id="item0" class="colorBox"></div>
                    <div id="item1" class="colorBox"></div>
                    <div id="item2" class="colorBox"></div>
                    <div id="item3" class="colorBox"></div>
                    <div id="item4" class="colorBox"></div>
                    <div id="item5" class="colorBox"></div>
                </div>
            </div>
            <div id="inheritSnapCoordinate" class="horizontalGallery inheritSnapCoordinate" style="-webkit-scroll-snap-coordinate: inherit;">
                <div class="horizontalGalleryDrawer">
                    <div id="item0" class="colorBox"></div>
                    <div id="item1" class="colorBox"></div>
                    <div id="item2" class="colorBox"></div>
                    <div id="item3" class="colorBox"></div>
                    <div id="item4" class="colorBox"></div>
                    <div id="item5" class="colorBox"></div>
                </div>
            </div>
            <div id="inheritAll" class="horizontalGallery inheritAll" style="-webkit-scroll-snap-type: inherit; -webkit-scroll-snap-points-x: inherit; -webkit-scroll-snap-points-y: inherit; -webkit-scroll-snap-destination: inherit; -webkit-scroll-snap-coordinate: inherit;">
                <div class="horizontalGalleryDrawer">
                    <div id="item0" class="colorBox"></div>
                    <div id="item1" class="colorBox"></div>
                    <div id="item2" class="colorBox"></div>
                    <div id="item3" class="colorBox"></div>
                    <div id="item4" class="colorBox"></div>
                    <div id="item5" class="colorBox"></div>
                </div>
            </div>
            <div id="inheritNone" class="horizontalGallery inheritNone">
                <div class="horizontalGalleryDrawer">
                    <div id="item0" class="colorBox"></div>
                    <div id="item1" class="colorBox"></div>
                    <div id="item2" class="colorBox"></div>
                    <div id="item3" class="colorBox"></div>
                    <div id="item4" class="colorBox"></div>
                    <div id="item5" class="colorBox"></div>
                </div>
            </div>
        </div>
        <div id="console"></div>
        <script>
        description("Tests 'inherit' on scroll snap properties.");

        var noInherit = document.getElementById('noInherit');
        shouldBeEmptyString("noInherit.style['-webkit-scroll-snap-type']");
        shouldBeEqualToString("window.getComputedStyle(noInherit).getPropertyValue('-webkit-scroll-snap-type')", "mandatory");
        shouldBeEmptyString("noInherit.style['-webkit-scroll-snap-points-x']");
        shouldBeEqualToString("window.getComputedStyle(noInherit).getPropertyValue('-webkit-scroll-snap-points-x')", "repeat(100%)");
        shouldBeEmptyString("noInherit.style['-webkit-scroll-snap-points-y']");
        shouldBeEqualToString("window.getComputedStyle(noInherit).getPropertyValue('-webkit-scroll-snap-points-y')", "repeat(100%)");
        shouldBeEmptyString("noInherit.style['-webkit-scroll-snap-destination']");
        shouldBeEqualToString("window.getComputedStyle(noInherit).getPropertyValue('-webkit-scroll-snap-destination')", "10% 11%");
        shouldBeEmptyString("noInherit.style['-webkit-scroll-snap-coordinate']");
        shouldBeEqualToString("window.getComputedStyle(noInherit).getPropertyValue('-webkit-scroll-snap-coordinate')", "none");

        var inheritType = document.getElementById('inheritType');
        shouldBeEqualToString("inheritType.style['-webkit-scroll-snap-type']", "inherit");
        shouldBeEqualToString("window.getComputedStyle(inheritType).getPropertyValue('-webkit-scroll-snap-type')", "proximity");
        shouldBeEmptyString("inheritType.style['-webkit-scroll-snap-points-x']");
        shouldBeEqualToString("window.getComputedStyle(inheritType).getPropertyValue('-webkit-scroll-snap-points-x')", "repeat(100%)");
        shouldBeEmptyString("inheritType.style['-webkit-scroll-snap-points-y']");
        shouldBeEqualToString("window.getComputedStyle(inheritType).getPropertyValue('-webkit-scroll-snap-points-y')", "repeat(100%)");
        shouldBeEmptyString("inheritType.style['-webkit-scroll-snap-destination']");
        shouldBeEqualToString("window.getComputedStyle(inheritType).getPropertyValue('-webkit-scroll-snap-destination')", "0px 0px");
        shouldBeEmptyString("inheritType.style['-webkit-scroll-snap-coordinate']");
        shouldBeEqualToString("window.getComputedStyle(inheritType).getPropertyValue('-webkit-scroll-snap-coordinate')", "15px 122px");

        var inheritXPoints = document.getElementById('inheritXPoints');
        shouldBeEmptyString("inheritXPoints.style['-webkit-scroll-snap-type']");
        shouldBeEqualToString("window.getComputedStyle(inheritXPoints).getPropertyValue('-webkit-scroll-snap-type')", "mandatory");
        shouldBeEqualToString("inheritXPoints.style['-webkit-scroll-snap-points-x']", "inherit");
        shouldBeEqualToString("window.getComputedStyle(inheritXPoints).getPropertyValue('-webkit-scroll-snap-points-x')", "repeat(200%)");
        shouldBeEmptyString("inheritXPoints.style['-webkit-scroll-snap-points-y']");
        shouldBeEqualToString("window.getComputedStyle(inheritXPoints).getPropertyValue('-webkit-scroll-snap-points-y')", "repeat(100%)");
        shouldBeEmptyString("inheritXPoints.style['-webkit-scroll-snap-destination']");
        shouldBeEqualToString("window.getComputedStyle(inheritXPoints).getPropertyValue('-webkit-scroll-snap-destination')", "50% 70%");
        shouldBeEmptyString("inheritXPoints.style['-webkit-scroll-snap-coordinate']");
        shouldBeEqualToString("window.getComputedStyle(inheritXPoints).getPropertyValue('-webkit-scroll-snap-coordinate')", "100% 50%");

        var inheritYPoints = document.getElementById('inheritYPoints');
        shouldBeEmptyString("inheritYPoints.style['-webkit-scroll-snap-type']");
        shouldBeEqualToString("window.getComputedStyle(inheritYPoints).getPropertyValue('-webkit-scroll-snap-type')", "mandatory");
        shouldBeEmptyString("inheritYPoints.style['-webkit-scroll-snap-points-x']");
        shouldBeEqualToString("window.getComputedStyle(inheritYPoints).getPropertyValue('-webkit-scroll-snap-points-x')", "repeat(100%)");
        shouldBeEqualToString("inheritYPoints.style['-webkit-scroll-snap-points-y']", "inherit");
        shouldBeEqualToString("window.getComputedStyle(inheritYPoints).getPropertyValue('-webkit-scroll-snap-points-y')", "repeat(200%)");
        shouldBeEmptyString("inheritYPoints.style['-webkit-scroll-snap-destination']");
        shouldBeEqualToString("window.getComputedStyle(inheritYPoints).getPropertyValue('-webkit-scroll-snap-destination')", "10px 20px");
        shouldBeEmptyString("inheritYPoints.style['-webkit-scroll-snap-coordinate']");
        shouldBeEqualToString("window.getComputedStyle(inheritYPoints).getPropertyValue('-webkit-scroll-snap-coordinate')", "50% 100%");

        var inheritDestination = document.getElementById('inheritDestination');
        shouldBeEmptyString("inheritDestination.style['-webkit-scroll-snap-type']");
        shouldBeEqualToString("window.getComputedStyle(inheritDestination).getPropertyValue('-webkit-scroll-snap-type')", "proximity");
        shouldBeEmptyString("inheritDestination.style['-webkit-scroll-snap-points-x']");
        shouldBeEqualToString("window.getComputedStyle(inheritDestination).getPropertyValue('-webkit-scroll-snap-points-x')", "repeat(100%)");
        shouldBeEmptyString("inheritDestination.style['-webkit-scroll-snap-points-y']");
        shouldBeEqualToString("window.getComputedStyle(inheritDestination).getPropertyValue('-webkit-scroll-snap-points-y')", "repeat(100%)");
        shouldBeEqualToString("inheritDestination.style['-webkit-scroll-snap-destination']", "inherit");
        shouldBeEqualToString("window.getComputedStyle(inheritDestination).getPropertyValue('-webkit-scroll-snap-destination')", "25% 25%");
        shouldBeEmptyString("inheritDestination.style['-webkit-scroll-snap-coordinate']");
        shouldBeEqualToString("window.getComputedStyle(inheritDestination).getPropertyValue('-webkit-scroll-snap-coordinate')", "10px 10px");

        var inheritSnapCoordinate = document.getElementById('inheritSnapCoordinate');
        shouldBeEmptyString("inheritSnapCoordinate.style['-webkit-scroll-snap-type']");
        shouldBeEqualToString("window.getComputedStyle(inheritSnapCoordinate).getPropertyValue('-webkit-scroll-snap-type')", "mandatory");
        shouldBeEmptyString("inheritSnapCoordinate.style['-webkit-scroll-snap-points-x']");
        shouldBeEqualToString("window.getComputedStyle(inheritSnapCoordinate).getPropertyValue('-webkit-scroll-snap-points-x')", "repeat(100%)");
        shouldBeEmptyString("inheritSnapCoordinate.style['-webkit-scroll-snap-points-y']");
        shouldBeEqualToString("window.getComputedStyle(inheritSnapCoordinate).getPropertyValue('-webkit-scroll-snap-points-y')", "repeat(100%)");
        shouldBeEmptyString("inheritSnapCoordinate.style['-webkit-scroll-snap-destination']");
        shouldBeEqualToString("window.getComputedStyle(inheritSnapCoordinate).getPropertyValue('-webkit-scroll-snap-destination')", "10% 10%");
        shouldBeEqualToString("inheritSnapCoordinate.style['-webkit-scroll-snap-coordinate']", "inherit");
        shouldBeEqualToString("window.getComputedStyle(inheritSnapCoordinate).getPropertyValue('-webkit-scroll-snap-coordinate')", "20px 20px");

        var inheritAll = document.getElementById('inheritAll');
        shouldBeEqualToString("inheritAll.style['-webkit-scroll-snap-type']", "inherit");
        shouldBeEqualToString("window.getComputedStyle(inheritAll).getPropertyValue('-webkit-scroll-snap-type')", "proximity");
        shouldBeEqualToString("inheritAll.style['-webkit-scroll-snap-points-x']", "inherit");
        shouldBeEqualToString("window.getComputedStyle(inheritAll).getPropertyValue('-webkit-scroll-snap-points-x')", "repeat(200%)");
        shouldBeEqualToString("inheritAll.style['-webkit-scroll-snap-points-y']", "inherit");
        shouldBeEqualToString("window.getComputedStyle(inheritAll).getPropertyValue('-webkit-scroll-snap-points-y')", "repeat(200%)");
        shouldBeEqualToString("inheritAll.style['-webkit-scroll-snap-destination']", "inherit");
        shouldBeEqualToString("window.getComputedStyle(inheritAll).getPropertyValue('-webkit-scroll-snap-destination')", "25% 25%");
        shouldBeEqualToString("inheritAll.style['-webkit-scroll-snap-coordinate']", "inherit");
        shouldBeEqualToString("window.getComputedStyle(inheritAll).getPropertyValue('-webkit-scroll-snap-coordinate')", "20px 20px");

        var inheritNone = document.getElementById('inheritNone');
        shouldBeEmptyString("inheritNone.style['-webkit-scroll-snap-type']");
        shouldBeEqualToString("window.getComputedStyle(inheritNone).getPropertyValue('-webkit-scroll-snap-type')", "none");
        shouldBeEmptyString("inheritNone.style['-webkit-scroll-snap-points-x']");
        shouldBeEqualToString("window.getComputedStyle(inheritNone).getPropertyValue('-webkit-scroll-snap-points-x')", "none");
        shouldBeEmptyString("inheritNone.style['-webkit-scroll-snap-points-y']");
        shouldBeEqualToString("window.getComputedStyle(inheritNone).getPropertyValue('-webkit-scroll-snap-points-y')", "none");
        shouldBeEmptyString("inheritNone.style['-webkit-scroll-snap-destination']");
        shouldBeEqualToString("window.getComputedStyle(inheritNone).getPropertyValue('-webkit-scroll-snap-destination')", "0px 0px");
        shouldBeEmptyString("inheritNone.style['-webkit-scroll-snap-coordinate']");
        shouldBeEqualToString("window.getComputedStyle(inheritNone).getPropertyValue('-webkit-scroll-snap-coordinate')", "none");
        </script>
        <script src="../../resources/js-test-post.js"></script>
    </body>
</html>
