LayoutTests/css3/scroll-snap/scroll-snap-inherit-expected.txt

Tests 'inherit' on scroll snap properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS noInherit.style['-webkit-scroll-snap-type'] is ""
PASS window.getComputedStyle(noInherit).getPropertyValue('-webkit-scroll-snap-type') is "mandatory"
PASS noInherit.style['-webkit-scroll-snap-points-x'] is ""
PASS window.getComputedStyle(noInherit).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(100%)"
PASS noInherit.style['-webkit-scroll-snap-points-y'] is ""
PASS window.getComputedStyle(noInherit).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(100%)"
PASS noInherit.style['-webkit-scroll-snap-destination'] is ""
PASS window.getComputedStyle(noInherit).getPropertyValue('-webkit-scroll-snap-destination') is "10% 11%"
PASS noInherit.style['-webkit-scroll-snap-coordinate'] is ""
PASS window.getComputedStyle(noInherit).getPropertyValue('-webkit-scroll-snap-coordinate') is "none"
PASS inheritType.style['-webkit-scroll-snap-type'] is "inherit"
PASS window.getComputedStyle(inheritType).getPropertyValue('-webkit-scroll-snap-type') is "proximity"
PASS inheritType.style['-webkit-scroll-snap-points-x'] is ""
PASS window.getComputedStyle(inheritType).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(100%)"
PASS inheritType.style['-webkit-scroll-snap-points-y'] is ""
PASS window.getComputedStyle(inheritType).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(100%)"
PASS inheritType.style['-webkit-scroll-snap-destination'] is ""
PASS window.getComputedStyle(inheritType).getPropertyValue('-webkit-scroll-snap-destination') is "0px 0px"
PASS inheritType.style['-webkit-scroll-snap-coordinate'] is ""
PASS window.getComputedStyle(inheritType).getPropertyValue('-webkit-scroll-snap-coordinate') is "15px 122px"
PASS inheritXPoints.style['-webkit-scroll-snap-type'] is ""
PASS window.getComputedStyle(inheritXPoints).getPropertyValue('-webkit-scroll-snap-type') is "mandatory"
PASS inheritXPoints.style['-webkit-scroll-snap-points-x'] is "inherit"
PASS window.getComputedStyle(inheritXPoints).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(200%)"
PASS inheritXPoints.style['-webkit-scroll-snap-points-y'] is ""
PASS window.getComputedStyle(inheritXPoints).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(100%)"
PASS inheritXPoints.style['-webkit-scroll-snap-destination'] is ""
PASS window.getComputedStyle(inheritXPoints).getPropertyValue('-webkit-scroll-snap-destination') is "50% 70%"
PASS inheritXPoints.style['-webkit-scroll-snap-coordinate'] is ""
PASS window.getComputedStyle(inheritXPoints).getPropertyValue('-webkit-scroll-snap-coordinate') is "100% 50%"
PASS inheritYPoints.style['-webkit-scroll-snap-type'] is ""
PASS window.getComputedStyle(inheritYPoints).getPropertyValue('-webkit-scroll-snap-type') is "mandatory"
PASS inheritYPoints.style['-webkit-scroll-snap-points-x'] is ""
PASS window.getComputedStyle(inheritYPoints).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(100%)"
PASS inheritYPoints.style['-webkit-scroll-snap-points-y'] is "inherit"
PASS window.getComputedStyle(inheritYPoints).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(200%)"
PASS inheritYPoints.style['-webkit-scroll-snap-destination'] is ""
PASS window.getComputedStyle(inheritYPoints).getPropertyValue('-webkit-scroll-snap-destination') is "10px 20px"
PASS inheritYPoints.style['-webkit-scroll-snap-coordinate'] is ""
PASS window.getComputedStyle(inheritYPoints).getPropertyValue('-webkit-scroll-snap-coordinate') is "50% 100%"
PASS inheritDestination.style['-webkit-scroll-snap-type'] is ""
PASS window.getComputedStyle(inheritDestination).getPropertyValue('-webkit-scroll-snap-type') is "proximity"
PASS inheritDestination.style['-webkit-scroll-snap-points-x'] is ""
PASS window.getComputedStyle(inheritDestination).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(100%)"
PASS inheritDestination.style['-webkit-scroll-snap-points-y'] is ""
PASS window.getComputedStyle(inheritDestination).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(100%)"
PASS inheritDestination.style['-webkit-scroll-snap-destination'] is "inherit"
PASS window.getComputedStyle(inheritDestination).getPropertyValue('-webkit-scroll-snap-destination') is "25% 25%"
PASS inheritDestination.style['-webkit-scroll-snap-coordinate'] is ""
PASS window.getComputedStyle(inheritDestination).getPropertyValue('-webkit-scroll-snap-coordinate') is "10px 10px"
PASS inheritSnapCoordinate.style['-webkit-scroll-snap-type'] is ""
PASS window.getComputedStyle(inheritSnapCoordinate).getPropertyValue('-webkit-scroll-snap-type') is "mandatory"
PASS inheritSnapCoordinate.style['-webkit-scroll-snap-points-x'] is ""
PASS window.getComputedStyle(inheritSnapCoordinate).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(100%)"
PASS inheritSnapCoordinate.style['-webkit-scroll-snap-points-y'] is ""
PASS window.getComputedStyle(inheritSnapCoordinate).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(100%)"
PASS inheritSnapCoordinate.style['-webkit-scroll-snap-destination'] is ""
PASS window.getComputedStyle(inheritSnapCoordinate).getPropertyValue('-webkit-scroll-snap-destination') is "10% 10%"
PASS inheritSnapCoordinate.style['-webkit-scroll-snap-coordinate'] is "inherit"
PASS window.getComputedStyle(inheritSnapCoordinate).getPropertyValue('-webkit-scroll-snap-coordinate') is "20px 20px"
PASS inheritAll.style['-webkit-scroll-snap-type'] is "inherit"
PASS window.getComputedStyle(inheritAll).getPropertyValue('-webkit-scroll-snap-type') is "proximity"
PASS inheritAll.style['-webkit-scroll-snap-points-x'] is "inherit"
PASS window.getComputedStyle(inheritAll).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(200%)"
PASS inheritAll.style['-webkit-scroll-snap-points-y'] is "inherit"
PASS window.getComputedStyle(inheritAll).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(200%)"
PASS inheritAll.style['-webkit-scroll-snap-destination'] is "inherit"
PASS window.getComputedStyle(inheritAll).getPropertyValue('-webkit-scroll-snap-destination') is "25% 25%"
PASS inheritAll.style['-webkit-scroll-snap-coordinate'] is "inherit"
PASS window.getComputedStyle(inheritAll).getPropertyValue('-webkit-scroll-snap-coordinate') is "20px 20px"
PASS inheritNone.style['-webkit-scroll-snap-type'] is ""
PASS window.getComputedStyle(inheritNone).getPropertyValue('-webkit-scroll-snap-type') is "none"
PASS inheritNone.style['-webkit-scroll-snap-points-x'] is ""
PASS window.getComputedStyle(inheritNone).getPropertyValue('-webkit-scroll-snap-points-x') is "none"
PASS inheritNone.style['-webkit-scroll-snap-points-y'] is ""
PASS window.getComputedStyle(inheritNone).getPropertyValue('-webkit-scroll-snap-points-y') is "none"
PASS inheritNone.style['-webkit-scroll-snap-destination'] is ""
PASS window.getComputedStyle(inheritNone).getPropertyValue('-webkit-scroll-snap-destination') is "0px 0px"
PASS inheritNone.style['-webkit-scroll-snap-coordinate'] is ""
PASS window.getComputedStyle(inheritNone).getPropertyValue('-webkit-scroll-snap-coordinate') is "none"
PASS successfullyParsed is true

TEST COMPLETE
